Added a context_preprocessor that adds a base_url variable. Started to work in alternative graphs

+++ web/siteup_api/models.py
@@ -74,7 +74,7 @@ class CheckLog(models.Model):
         max_length=255
     )
 
-    response_time = models.IntegerField(
+    response_time = models.FloatField(
         default=0
     )
